Refactor paymentMethodType utility
