+++ Sources/PrimerSDK/Classes/Extensions & Utilities/PaymentMethodTypeProviding.swift
@@ -1,0 +1,24 @@
+//
+//  PaymentMethodTypeProviding.swift
+//  PrimerSDK
+//
+//  Created by Niall Quinn on 06/12/23.
+//
+
+import Foundation
+
+protocol PaymentMethodTypeViaPaymentMethodTokenDataProviding {
+    var paymentMethodTokenData: PrimerPaymentMethodTokenData? { get }
+    var paymentMethodType: String { get }
+}
+
+extension PaymentMethodTypeViaPaymentMethodTokenDataProviding {
+    var paymentMethodType: String {
+        if let pmt = paymentMethodTokenData?.paymentMethodType {
+            return pmt
+        } else if let pmt = paymentMethodTokenData?.paymentInstrumentData?.paymentMethodType {
+            return pmt
+        }
+        return "UNKNOWN"
+    }
+}
